+++ ranch-account/src/main/java/org/lpw/ranch/account/type/ConsumeImpl.java
@@ -18,7 +18,7 @@ public class ConsumeImpl extends AccountTypeSupport implements AccountType {
     }
 
     @Override
-    public String change(AccountModel account, String channel, int amount, Map<String, String> map) {
+    public String change(AccountModel account, String channel, long amount, Map<String, String> map) {
         return out(account, channel, amount, map);
     }
